xmhf-hwm: cpu: fix SGDT emulation; eliminates out of bound errors with gettssbase within libxmhfhw. successful verification of s5_setupcpustate.

+++ xmhf-hwm/xmhfhwm_cpu.c
@@ -394,7 +394,7 @@ void _impl_xmhfhwm_cpu_insn_sgdt_mesp(int index){
 	u32 *tmem_gdtbase;
 	u16 *tmem_gdtlimit;
 	tmem_gdtlimit = (u16 *)((u32)((int)xmhfhwm_cpu_gprs_esp + (int)index));
-	tmem_gdtbase = (u32 *)((u32)((int)xmhfhwm_cpu_gprs_esp + (int)index) + sizeof(u32));
+	tmem_gdtbase = (u32 *)((u32)((int)xmhfhwm_cpu_gprs_esp + (int)index) + sizeof(u16));
 	*tmem_gdtlimit = xmhfhwm_cpu_gdtr_limit;
 	*tmem_gdtbase = xmhfhwm_cpu_gdtr_base;
 }
